Adicionando funcionalidade de cadastrar conta/ fazendo modificações para ela funcionar

+++ sistema_bancario/src/main/java/org/example/Conta.java
@@ -10,7 +10,7 @@ public class Conta {
     }
 
 
-    public Conta(long numeroConta, float saldo){
+    public Conta(long numeroConta, double saldo){
         this.numeroConta = numeroConta;
         this.saldo = saldo;
     }
